+++ bootique-shiro-jdbc/src/main/java/io/bootique/shiro/jdbc/ShiroJdbcModule.java
@@ -1,0 +1,12 @@
+package io.bootique.shiro.jdbc;
+
+import com.google.inject.Binder;
+import com.google.inject.Module;
+
+public class ShiroJdbcModule implements Module {
+
+    @Override
+    public void configure(Binder binder) {
+
+    }
+}