+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 0.1
 
 * #1 Basic implementation
+* #2 Configurable JDBC Realm
 * #3 Generic mechanism for loading Realms from configuration and DI
 * #4 IniRealmFactory